feat: starting to manage cluster upgrades in the operator
whenever a new installation object pops up in the cluster and it points
to a different version than the current version we create a new plan and
delegate to autopilot the upgrade. for now this isn't working yet due to
a problem in upstream autopilot.

the idea here is that kots will create this installation object in the
cluster and then just observe the installation object until the upgrade
has been finished or failed.