bug: fix missing artifact copy job and autopilot airgap command
when upgrading a cluster from version X to version X we still need to
sync the images.tar and the helm charts as those may differ.

helm charts and images do not differ based on the embedded cluster
version but are based on the published kots application instead.

this commit introduces a few changes:

- we only sync up k0s versions if more than one installation exist.
- when upgrading we determine what is necessary to upgrade.
  - it might be that we only want to upgrade the k0s version.
  - it might be that we only want to sync the images tar file.
  - it might be that we want to sync both things.
  - it might be that nothing needs to be synced (already up to date).